Fix link
In OwnCloud, the result was that there was a leading '/' before the link (eg. "/public/webdaf/FILENAME"). In getOwncloudFiles.sh, another / is added at the start of the link. The resulting link will be like this: SERVER//public.php/webdav/FILENAME. This results in an error. To prevent this, the leading '/' needs to be omitted.